backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI/ML APIs
openai==1.10.0
anthropic==0.18.1

# Text processing and RAG
pymupdf==1.23.21
python-docx==1.1.0
sentence-transformers==2.3.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging and monitoring
loguru==0.7.2
